feat: separate file for SnowflakeSchemachangeSession
